Unet_buildings_detection/requirements.txt
python 3.6
pytorch
scipy
gdal
cv2